pfc_shaping/requirements.txt
# PFC Shaping 15min — FMV SA
# Python >= 3.10

# Data & calcul
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyarrow>=14.0.0      # Parquet I/O

# Machine Learning
scikit-learn>=1.3.0  # HuberRegressor, LinearRegression

# Lissage gaussien
# (scipy.ndimage inclus dans scipy)

# SMARD.de API (Bundesnetzagentur) — source primaire prix EPEX
requests>=2.28.0

# ENTSO-E Transparency API (fallback prix + source load/generation)
entsoe-py>=0.6.0
python-dotenv>=1.0.0

# Databricks SQL Connector (forwards EEX depuis EULER)
databricks-sql-connector>=3.0.0

# Calendrier fériés CH/DE
holidays>=0.40

# Configuration
PyYAML>=6.0
duckdb>=1.1.0

# Tests
pytest>=7.0.0

# Optionnel : visualisation (backtest / diagnostic)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Optionnel : planificateur (alternative à cron)
# APScheduler>=3.10.0
